rpm: replace old parser with kaitai struct based parser

+++ src/parsers/archivers/rpm/UnpackParser.py
@@ -20,35 +20,33 @@
 # version 3
 # SPDX-License-Identifier: AGPL-3.0-only
 
-'''
-Parse and unpack RPM files.
-'''
-
 
 import os
-from UnpackParser import WrappedUnpackParser
-from bangunpack import unpack_rpm
+import pathlib
+import shutil
+import tempfile
+
 from bangunpack import unpack_gzip
 from bangunpack import unpack_bzip2
 from bangunpack import unpack_xz
 from bangunpack import unpack_lzma
 from bangunpack import unpack_zstd
 
+from parsers.archivers.cpio import UnpackParser as cpio_unpack
+
+from FileResult import FileResult
+
 from UnpackParser import UnpackParser, check_condition
 from UnpackParserException import UnpackParserException
 from kaitaistruct import ValidationFailedError
 from . import rpm
 
-#class RpmUnpackParser(UnpackParser):
-class RpmUnpackParser(WrappedUnpackParser):
+class RpmUnpackParser(UnpackParser):
     extensions = []
     signatures = [
         (0, b'\xed\xab\xee\xdb')
     ]
     pretty_name = 'rpm'
-
-    def unpack_function(self, fileresult, scan_environment, offset, unpack_dir):
-        return unpack_rpm(fileresult, scan_environment, offset, unpack_dir)
 
     def parse(self):
         file_size = self.fileresult.filename.stat().st_size
@@ -61,31 +59,22 @@ class RpmUnpackParser(WrappedUnpackParser):
                         self.data.lead.type == rpm.Rpm.RpmTypes.source,
                         "invalid RPM type")
 
-        # extract the header + payload size, calculate the payload size
-        rpmsize = 0
-        for i in self.data.signature.index_records:
-            if i.tag == rpm.Rpm.SignatureTags.size:
-                rpmsize = i.body.values[0]
-        self.payload_size = rpmsize - self.data.header.header_size
-        check_condition(self.data.payload_offset + self.payload_size <= file_size,
-                        "payload cannot be outside of file")
-
         self.compressor_seen = False
         self.payload_format = ''
         # at most one compressor can be defined
         for i in self.data.header.index_records:
-            if i.tag == rpm.Rpm.HeaderTags.payload_compressor:
+            if i.header_tag == rpm.Rpm.HeaderTags.payload_compressor:
                 check_condition(not self.compressor_seen, "duplicate compressor defined")
                 self.compressor_seen = True
-            if i.tag == rpm.Rpm.HeaderTags.payload_format:
+            if i.header_tag == rpm.Rpm.HeaderTags.payload_format:
                 check_condition(self.payload_format == '', "duplicate compressor defined")
                 self.payload_format = i.body.values[0]
 
 
     def unpack(self):
         unpacked_files = []
-        if self.payload_format != 'cpio':
-            return(unpacked_files)
+        if self.payload_format not in ['cpio', 'drpm']:
+            return unpacked_files
 
         # then unpack the file. This depends on the compressor and the
         # payload format.  The default compressor is either gzip or XZ
@@ -96,54 +85,74 @@ class RpmUnpackParser(WrappedUnpackParser):
             compressor = 'gzip'
         else:
             for i in self.data.header.index_records:
-                if i.tag == rpm.Rpm.HeaderTags.payload_compressor:
+                if i.header_tag == rpm.Rpm.HeaderTags.payload_compressor:
                     compressor = i.body.values[0]
                     break
-        offset = self.offset + self.data.payload_offset
+
+        # write the payload to a temporary file first
+        temporary_file = tempfile.mkstemp(dir=self.scan_environment.temporarydirectory)
+        os.write(temporary_file[0], self.data.payload)
+        os.fdopen(temporary_file[0]).close()
+
+        fr = FileResult(None, temporary_file[1], set([]))
+        fr.set_filesize(len(self.data.payload))
+
         if compressor == 'gzip':
-            unpackresult = unpack_gzip(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_gzip(fr, self.scan_environment, 0, self.rel_unpack_dir)
         elif compressor == 'bzip2':
-            unpackresult = unpack_bzip2(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_bzip2(fr, self.scan_environment, 0, self.rel_unpack_dir)
         elif compressor == 'xz':
-            unpackresult = unpack_xz(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_xz(fr, self.scan_environment, 0, self.rel_unpack_dir)
         elif compressor == 'lzma':
-            unpackresult = unpack_lzma(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_lzma(fr, self.scan_environment, 0, self.rel_unpack_dir)
         elif compressor == 'zstd':
-            unpackresult = unpack_zstd(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_zstd(fr, self.scan_environment, 0, self.rel_unpack_dir)
         else:
             # gzip is default
-            unpackresult = unpack_gzip(self.fileresult, self.scan_environment, offset, self.rel_unpack_dir)
+            unpackresult = unpack_gzip(fr, self.scan_environment, 0, self.rel_unpack_dir)
+        os.unlink(temporary_file[1])
 
-        payloadfile = rpmunpackfiles[0][0]
-        payloadfile_full = scanenvironment.unpack_path(payloadfile)
+        payloadfile = unpackresult['filesandlabels'][0][0]
+        payloadfile_full = self.scan_environment.unpack_path(payloadfile)
 
-        # first move the payload file to a different location
-        # to avoid any potential name clashes
-        payloadsize = payloadfile_full.stat().st_size
-        payloaddir = pathlib.Path(tempfile.mkdtemp(dir=scanenvironment.temporarydirectory))
-        shutil.move(str(payloadfile_full), payloaddir)
+        if self.payload_format == 'drpm':
+            fr = FileResult(self.fileresult, self.rel_unpack_dir / os.path.basename(payloadfile), set())
+            unpacked_files.append(fr)
+        else:
+            # first move the payload file to a different location
+            # to avoid any potential name clashes
+            payloadsize = payloadfile_full.stat().st_size
+            payloaddir = pathlib.Path(tempfile.mkdtemp(dir=self.scan_environment.temporarydirectory))
+            shutil.move(str(payloadfile_full), payloaddir)
 
-        # create a file result object and pass it to the CPIO unpacker
-        fr = FileResult(fileresult,
-                payloaddir / os.path.basename(payloadfile),
-                set([]))
-        fr.set_filesize(payloadsize)
-        unpackresult = unpack_cpio(fr, scanenvironment, 0, unpackdir)
-        # cleanup
-        shutil.rmtree(payloaddir)
-        if not unpackresult['status']:
-            checkfile.close()
-            unpackingerror = {'offset': offset, 'fatal': False,
-                              'reason': 'could not unpack CPIO payload'}
-            return {'status': False, 'error': unpackingerror}
-        for i in unpackresult['filesandlabels']:
-            # TODO: is normpath necessary now that we use relative paths?
-            unpackedfilesandlabels.append((os.path.normpath(i[0]), i[1]))
+            # create a file result object and pass it to the CPIO unpacker
+            fr = FileResult(self.fileresult,
+                    payloaddir / os.path.basename(payloadfile),
+                    set([]))
+            fr.set_filesize(payloadsize)
 
-        return unpacked_files
+            # assuming that the CPIO data is always in "new ascii" format
+            cpio_parser = cpio_unpack.CpioNewAsciiUnpackParser(fr, self.scan_environment, self.rel_unpack_dir, 0)
+            try:
+                cpio_parser.open()
+                unpackresult = cpio_parser.parse_and_unpack()
+            except UnpackParserException as e:
+                raise UnpackParserException(e.args)
+            finally:
+                cpio_parser.close()
+
+            # walk the results and add them.
+            for i in unpackresult.unpacked_files:
+                i.parent_path = self.fileresult.filename
+                unpacked_files.append(i)
+
+            shutil.rmtree(payloaddir)
+
+        return(unpacked_files)
+
 
     def calculate_unpacked_size(self):
-        self.unpacked_size = self.data.payload_offset + self.payload_size
+        self.unpacked_size = self.data.ofs_payload + len(self.data.payload)
 
     def set_metadata_and_labels(self):
         """sets metadata and labels for the unpackresults"""
@@ -164,15 +173,14 @@ class RpmUnpackParser(WrappedUnpackParser):
         # store signature tags
         metadata['signature_tags'] = {}
         for i in self.data.signature.index_records:
-            metadata['signature_tags'][i.tag.value] = i.body.values
+            metadata['signature_tags'][i.signature_tag.value] = i.body.values
 
         # store header tags
         metadata['header_tags'] = {}
         for i in self.data.header.index_records:
-            metadata['header_tags'][i.tag.value] = i.body.values
+            metadata['header_tags'][i.header_tag.value] = i.body.values
 
         if self.payload_format == 'drpm':
-            payloadfile_rel = scanenvironment.rel_unpack_path(payloadfile)
             labels.append('delta rpm')
 
         self.unpack_results.set_metadata(metadata)
